# docs/api_surface/OpenAPI.yaml
openapi: 3.0.3
info:
  title: Campus Resource Hub API
  version: 1.0.0
  description: |
    RESTful API for the Campus Resource Hub - A university resource booking and management system.
    
    **Implementation Status:**
    - ✅ existing: Backend implemented and functional
    - ⚠️ partial: Partially implemented
    - ❌ missing: Not yet implemented
    
    **Authentication:** Session-based cookies (Flask-Login)
    
    **Base URL:** `/api`
  contact:
    name: Campus Resource Hub Team
    email: support@campusresourcehub.edu
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://production-domain.com/api
    description: Production server (TBD)
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask-Login session cookie for authentication
    
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF protection token required for all state-changing operations (POST, PUT, PATCH, DELETE).
        Obtain token from GET /api/auth/csrf-token endpoint.
        Token lifetime: 1 hour
        Alternative header name: X-CSRFToken
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
        thread_id:
          type: string
        sender_id:
          type: integer
        receiver_id:
          type: integer
        content:
          type: string
        is_read:
          type: boolean
        timestamp:
          type: string
          format: date-time
      required: [id, sender_id, receiver_id, content]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
  responses:
...
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per hour
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests (will be 0)
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds until retry allowed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Too Many Requests"
              message:
                type: string
                example: "Rate limit exceeded. Please try again later."
              retry_after:
                type: string
                example: "60 seconds"
              status:
                type: integer
                example: 429
paths:
...
  /messages:
    get:
      tags: [Messages]
      summary: List message threads
      x-status: existing
      security:
        - SessionCookie: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of threads
          content:
            application/json:
              schema:
                type: object
                properties:
                  threads:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    
    post:
      tags: [Messages]
      summary: Send message
      x-status: existing
      description: |
        Send a message to another user.
        Rate limit: 30 per hour
      security:
        - SessionCookie: []
        - CSRFToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receiver_id, content]
              properties:
                receiver_id:
                  type: integer
                content:
                  type: string
                resource_id:
                  type: integer
      responses:
        '201':
          description: Message sent
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
        '429':
          $ref: '#/components/responses/RateLimitError'
  /messages/thread/{threadId}:
    get:
      tags: [Messages]
      summary: Get thread messages
      x-status: existing
      security:
        - SessionCookie: []
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Thread messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
  /messages/unread-count:
    get:
      tags: [Messages]
      summary: Get unread message count
      x-status: existing
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  unread_count:
                    type: integer
...